services/recommendation-engine/requirements.txt
boto3==1.28.25
psycopg2-binary==2.9.7
python-dotenv==1.0.0
